pyproject.toml
[tool.poetry]
name = "organize"
version = "0.1.0"
description = ""
authors = ["Pierre-Yves Gillier <github@pygillier.me>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
django-vite = "^2.0.2"
graphene-django = "==3.0.0b7"
graphql-core = "==3.1.0"
django-graphql-jwt = "^0.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
